backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-google-genai
langchain-pinecone
pinecone-client
pypdf2
python-dotenv
pydantic
supabase
tiktoken
sentence-transformers
chromadb
faiss-cpu 